+++ src/realm/sync/client.hpp
@@ -12,6 +12,7 @@
 #include <realm/util/buffer.hpp>
 #include <realm/util/functional.hpp>
 #include <realm/sync/client_base.hpp>
+#include <realm/sync/socket_provider.hpp>
 #include <realm/sync/subscriptions.hpp>
 
 namespace realm::sync {
@@ -37,10 +38,13 @@ public:
     Client(Client&&) noexcept;
     ~Client() noexcept;
 
+    /// Used by testing to post a function handler onto the event loop
+    void post_for_testing(SyncSocketProvider::FunctionHandler&& handler);
+
     /// Run the internal event-loop of the client. At most one thread may
     /// execute run() at any given time. The call will not return until somebody
     /// calls stop().
-    void run();
+    void run() noexcept;
 
     /// See run().
     ///