+++ test/sync_fixtures.hpp
@@ -547,7 +547,6 @@ public:
         }
 
         m_server_threads.resize(num_servers);
-        m_client_threads.resize(num_clients);
 
         m_simulated_server_error_rates.resize(num_servers);
         m_simulated_client_error_rates.resize(num_clients);
@@ -563,10 +562,6 @@ public:
     {
         unit_test::TestContext& test_context = m_test_context;
         stop();
-        for (int i = 0; i < m_num_clients; ++i) {
-            if (m_client_threads[i].joinable())
-                CHECK(!m_client_threads[i].join());
-        }
         for (int i = 0; i < m_num_servers; ++i) {
             if (m_server_threads[i].joinable())
                 CHECK(!m_server_threads[i].join());
@@ -594,15 +589,26 @@ public:
         m_connection_state_change_listeners[client_index] = std::move(handler_2);
     }
 
-    // Must be called before start().
     void set_client_side_error_rate(int client_index, int n, int m)
     {
-        m_simulated_client_error_rates[client_index] = std::make_pair(n, m);
+        REALM_ASSERT(client_index >= 0 && client_index < m_num_clients);
+        auto sim = std::make_pair(n, m);
+        // Save the simulated error rate
+        m_simulated_client_error_rates[client_index] = sim;
+
+        // Post the new simulated error rate
+        using sf = _impl::SimulatedFailure;
+        // Post it onto the event loop to update the event loop thread
+        m_clients[client_index]->post_for_testing([sim = std::move(sim)](Status) {
+            sf::prime_random(sf::sync_client__read_head, sim.first, sim.second,
+                             random_int<uint_fast64_t>()); // Seed from global generator
+        });
     }
 
     // Must be called before start().
     void set_server_side_error_rate(int server_index, int n, int m)
     {
+        REALM_ASSERT(server_index >= 0 && server_index < m_num_servers);
         m_simulated_server_error_rates[server_index] = std::make_pair(n, m);
     }
 
@@ -612,10 +618,6 @@ public:
             m_server_threads[i].start([this, i] {
                 run_server(i);
             });
-        for (int i = 0; i < m_num_clients; ++i)
-            m_client_threads[i].start([this, i] {
-                run_client(i);
-            });
     }
 
     // Use either the methods below or `start()`.
@@ -624,14 +626,6 @@ public:
         REALM_ASSERT(index >= 0 && index < m_num_servers);
         m_server_threads[index].start([this, index] {
             run_server(index);
-        });
-    }
-
-    void start_client(int index)
-    {
-        REALM_ASSERT(index >= 0 && index < m_num_clients);
-        m_client_threads[index].start([this, index] {
-            run_client(index);
         });
     }
 
@@ -648,16 +642,23 @@ public:
     void stop_client(int index)
     {
         REALM_ASSERT(index >= 0 && index < m_num_clients);
-        m_clients[index]->stop();
-        unit_test::TestContext& test_context = m_test_context;
-        if (m_client_threads[index].joinable())
-            CHECK(!m_client_threads[index].join());
+        auto& client = get_client(index);
+        auto sim = m_simulated_client_error_rates[index];
+        if (sim.first != 0) {
+            using sf = _impl::SimulatedFailure;
+            // If we're using a simulated failure, clear it by posting onto the event loop
+            client.post_for_testing([](Status) mutable {
+                sf::unprime(sf::sync_client__read_head); // Clear the sim failure set when started
+            });
+        }
+        // We can't wait for clearing the simulated failure since some tests stop the client early
+        client.stop();
     }
 
     void stop()
     {
         for (int i = 0; i < m_num_clients; ++i)
-            m_clients[i]->stop();
+            stop_client(i);
         for (int i = 0; i < m_num_servers; ++i)
             m_servers[i]->stop();
     }
@@ -796,7 +797,6 @@ private:
     std::vector<std::function<ConnectionStateChangeListener>> m_connection_state_change_listeners;
     std::vector<port_type> m_server_ports;
     std::vector<ThreadWrapper> m_server_threads;
-    std::vector<ThreadWrapper> m_client_threads;
     std::vector<std::pair<int, int>> m_simulated_server_error_rates;
     std::vector<std::pair<int, int>> m_simulated_client_error_rates;
     std::vector<uint_least64_t> m_allow_server_errors;
@@ -834,28 +834,6 @@ private:
             return;
         stop();
         m_server_loggers[i]->error("Exception was throw from server[%1]'s event loop", i + 1);
-    }
-
-    void run_client(int i)
-    {
-        auto do_run_client = [this, i] {
-            auto sim = m_simulated_client_error_rates[i];
-            if (sim.first != 0) {
-                using sf = _impl::SimulatedFailure;
-                sf::RandomPrimeGuard pg(sf::sync_client__read_head, sim.first, sim.second,
-                                        random_int<uint_fast64_t>()); // Seed from global generator
-                m_clients[i]->run();
-            }
-            else {
-                m_clients[i]->run();
-            }
-            m_clients[i]->stop();
-        };
-        unit_test::TestContext& test_context = m_test_context;
-        if (CHECK_NOTHROW(do_run_client()))
-            return;
-        stop();
-        m_server_loggers[i]->error("Exception was throw from client[%1]'s event loop", i + 1);
     }
 };